requirements.txt
python-telegram-bot==13.11
requests==2.27.*